requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
emoji==2.2.0
freeport==0.1.15
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
psycopg2-binary==2.9.6
pytz==2023.3
redis==4.5.4
